Ring constructor: ensure distinct primes, improve prime ranges
Loop until disctint primes are found. Can loop forever if that's
not satisfiable, but it's hard to check in advance for that.

This also improves the logic for the range of P, especially when
L is odd. The N range is 2^(L-1):2^L. This takes the sqrt of
each by dividing the exponent in half: 2^((L-1)/2):2^(L/2), except
that we want even powers of two so we round the lower one down and
the upper one up: 2^fld(L-1,2):2^cld(L,2).