Ring: check feasibility better, construct small instances faster
Someone paradoxically, it can be tricky to construct small instances of
rings since they are more constrained. Revamped logic does a better job
of detecting when the requested configuration is infeasible and finds a
feasible configuration more efficiently when it is.

The secret sauce here is narrowing the bounds on P and Q earlier and
more accurately. The new `narrow_prime_range` function takes input
bounds on P (with a scale factor) and finds the largest and smallest P
values that are actually feasible. If the interval becaomes empty then
we know that it cannot be done.

At the higher level, we use `narrow_prime_range` to detect infeasible
situations. We start with the minimum lower bound we want on both prime
factors; then we iterate back and forth between bounds on P and Q to
mutually narrow them down such that actual prime pairs exist. Last, we
check that one of the extremal solutions is usable in the sense that all
the factors are coprime. nce we're sure that a usable solution exists,
we can start sampling until we get some usable solution.