Client: introduce struct representing a client
This struct captures what the client knows and generates, allowing us to
simulate the client side of the protocol as well as the server side,
which is captured by the Ring structure.

In our simulation, the client picks a random semigenerator, but in real
usage the server would pick a single semigenerator and publish it for
all clients to use.

The client then picks a random Jacobi twist element, τ, which serves as
its master key value. It pre-computes a hash value of τ to salt resource
hashes with. We use SHA1 just because it's a bit faster and a convenient
size (128 bits), and cryptographic security is unnecessary here. Most of
the information in the hash output is subsequently fully randomized by
exponentiation, so the server only observes one HyperLogLog value of
actual information (≈ 18 bits). This is not nearly enough to hope to
try to do a preimage analysis on even a vaguely secure hash function.

For each request, the client does the following:

     h = hash_resource_class(client.salt, class)
     x = mod(τ * powermod(g, h, N), N)     # x = τ*g^h
     t = 2B*rand(rng, 0:(N-1)÷(2B)-1) + 1  # t ∈ ℤ_N st t = 1 mod 2B
     y = powermod(x, t, N)                 # y = x^t

On my system, this request-specific step takes about 1.6ms, which is
entirely acceptable from a performance overhead perspective.