Certifiable ring structure
This adds the ability to generate rings with certifiable modulus. It
also adds a RingCert type that captures all the information about a ring
that allows a client to check that it has a structure that cannot be
used to fingerprint clients. The Client type checks the certificate when
constructing itself from a cert object.