fix modsqrt errors for fix-sized integer types, add tests
In order to ensure that fixed-size integer overflow doesn't cause the
modsqrt function to return incorrect answers, we widen the type. This
doesn't change anything for BigInts but uses Int128 arithmetic for Int64
arguments and BigInt arithemtic for Int128 arguments.

Use powermod for checks since it handles this correctly. We could
potentially avoid widening by using powermod in modsqrt instead, but
since the real use case uses BigInts, it's not worth messing with.