prove N and λ(N) are coprime via simple Σ-protocol
Consider Z_N with N = P^2 Q with multiplicative structure:

     Z_N^* ≅ C_P × C_(P-1) × C_(Q-1)

If P divides Q-1 then there are two C_P components that can be used to
fingerprint clients. However, this kind of structure is ruled out by
showing that N is coprime to λ(N). This true iff x^N is bijective. We
could provide D = N^-1 mod λ(N) but then the client knows DN-1, which is
large multiple of λ(N). This might be fine, but we'd really rather not
give away a multiple of λ(N) at all, so instead we do a Σ-protocol to
show that we know the Nth roots of elements chosen by a cryptographic
hash. If P | Q-1 then gcd(N, λ(N)) == P. For random x ∈ ℤ_N, the chance
that there exists y such that y^N = x is 1/P. This is why we rule out
N being divisible by any p ≤ 1024: this lower bounds the chance that
a random x has an Nth root if gcd(N, λ(N)) ≠ 1 at 1/2^10. We only need
to prove the existence of three Nth roots to get down to a 1/2^30 chance
of passing the test by accident. And of course, if the prime factors are
actually large, then the chances are much, much slimmer.