Root test: provide roots for all elements with all Jacobi symbols
We were previously skipping elements with J(x) = -1, which cannot have a
square root. Instead, we can pick a fixed twist element τ and instead
check if τ*x has a root and if it does, provide it. This makes the
sample size all of the elements we look at, instead of a random subset.

There is also a chance that J(x) = 0, but in that case x still has a 50%
chance of having a square root, so we don't special case that. In a real
ring, of course, this is infinitessimally unlikely and if it were to
happen it would reveal the factorization of N to the client.