realized that Σ-protocol is unnecessary: can just provide Nth roots
If there was a way to do a ZKP that there exists D such that

     N*D = 1 mod λ(N)

that would be great, but the approach here is just to prove that enough
random hash-generated elements have Nth roots to make it very unlikely
that N isn't invertible mod λ(N). But we providing Nth roots of random
elements is harmless, so we don't need to hide them with a ZKP, we can
just privde the actual Nth roots in the cert.