refactor cert code; majorly simplify cert checking
This changes the certificate to simply include a vector of square roots
without the indices they correspond to. They are, however, presumed to
be in index order. The client can check them by searching for the next
hashed (non-twist) element that is equal to the square of the root. If
the client has to search past the SQRT_SAMPLES value, the test fails.
This means that incorrect square root values aren't explicitly detected.
Instead the client searches and never finds the corresponding square
so it eventually fails the test by exceeding the limit.

The cert only includes the mimimum necessary square root values. If it
includes more values than necessary, the client ignores them.