Mth root tests: proves that p ∤ λ(N) for p ≤ log2(N) unless p | B
The concern here is that if the multiplicative structure is

     ℤ_N^* ≅ C_2 × C_(2^m) × C_R

where R has many small prime factors, then information about x is leaked
by which of those small primes the R-exponent of x is not coprime with.
This is modulated by t, which can also be divisible by small primes, but
there is still potential information leakage. This test proves that no
primes ≤ log2(N) divide λ(N) except the ones we expect from B. There's
still a possibility of information leakage from larger prime factors
but it's ≤ 1 bit total, and very hard to extract, so we're ok with it.