Client: add noise to hash; avoids issue with zeros mod p and q
Previously, the ℤ_p and ℤ_q exponents could be zero, which ends up being
a potential fingerpint of a client. When N has correct structure this is
so astronomically unlikely that we may as well not bother, but then we
need to prove to the client that λ(N) doesn't have a bunch of smaller
prime factors that allow significant fingerprinting. Instead, this adds
noise to h that has no effect modulo B and 2^m, but which erases info
about divisibility in the rest of the ring.