Client: different approach to adding noise
This version is what I've used in the anonymity proof. It's a slightly
stronger construction since z is fully random and w = z^(B 2^m) so the
effectiveness of the noise does not depend on powers of g covering
enough of the ring. This is helpful since a malicious server could pick
a value of g that isn't actually a semigenerator.

Interestingly, this actually makes message generation _faster_. I think
the reason is that it avoids exponentiating by any full range values. In
the new code, exponentiation is done by `h`, which is small (128-bit),
`B*2^m``, which is also small (128-bit) and simple in the sense of few
bit set, and `t`, which is also small (128-bit). There are no longer any
large, random exponents. Instead there's a multiply by `w`, which is a
large random value, but that's much faster then exponentiation.

New timings: 80µs (on power) 135µs (on battery).