certificates: we don't need (N-1)/2 to be prime
Now that we add white noise and send wx^t instead of just x^t, we don't
need to show that gcd(P-1, Q-1) = 1, it's sufficient to show that B does
not divide N-1, which implies that gcd(B, P-1, Q-1) = 1. We also don't
need 2 to be a witness to the compositeness of N. Previously this was
virtually guaranteed because of primality of (N-1)/2. It no longer is
but we also don't need to check that N isn't prime. This makes any prime
with the (2 B p + 1)(2^m q + 1) structure certifiable.