certificates: use stronger unconditional square root test
This test is that when N has at most two distinct prime factors, for any
x, y in ℤ_N* with positive Jacobi symbol, at least one of {x, y, x*y} is
a quadratic residue. Since this is unconditional, the test is simpler
and requires far fewer samples to convince clients.