Client: more flexible logic for checking sufficient roots
This allows some flexibility in how many Nth roots the server provides.
Instead of a fixed lower count of Nth roots, it computes how large the
trial division limit, L, would have to be and has a maximum on L that
can be larger than the default L value. There's also a minium L value
and the client will always check for divisibility up to that value.

In particular, this allows a certificate produced by a server with one
ε value to be accepted by a client with a different ε value. There is
less flexibility in the square root logic since the fraction there is
fixed and 5/8 and there's no simple check that can reduce this.